requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
PyYAML
redis        # optional kalau mau caching
